+++ libft/includes/libft.h
@@ -13,6 +13,8 @@
 #ifndef LIBFT_H
 # define LIBFT_H
 
+#include <inttypes.h>
+
 # include <stdlib.h>
 # include <string.h>
 # include <unistd.h>
@@ -66,7 +68,7 @@ int				ft_strnequ(char const *s1, char const *s2, size_t n);
 char			*ft_strsub(char const *s, unsigned int start, size_t len);
 char			*ft_strjoin(char const *s1, char const *s2);
 char			*ft_strtrim(char const *s);
-char			*ft_itoa(int n);
+char			*ft_itoa(intmax_t n);
 char			**ft_strsplit(char const *s, char c);
 void			ft_putchar(char c);
 void			ft_putstr(char *str);
